requirements.txt
dashscope~=1.14.1
python-dotenv~=1.0.1
langchain~=0.1.6
pydantic~=2.6.1